クラス定義などを明確に分離
