+++ tools/yaml2x/yaml2rst/initializer.py
@@ -1,7 +1,8 @@
 import os
 import argparse
 import config
-from path import get_dest_dirnames, get_static_dest_files, get_src_path, TEMPLATE_DIR, TEMPLATE_FILENAMES
+from path import get_dest_dirnames, get_static_dest_files,  TEMPLATE_DIR, TEMPLATE_FILENAMES
+from a11y_guidelines import get_src_path
 from template_manager import TemplateManager
 
 def setup_parameters():
